Rename find to subscribe, extract finding code
